feat: implement adapted F/A
