+++ packages/semi-foundation/resizable/constants.ts
@@ -209,13 +209,4 @@ export const calculateNewMax = (
     };
 };
 export const normalizeToPair = <T>(val: T | [T, T]): [T, T] => (Array.isArray(val) ? val : [val, val]);
-export const isTouchEvent = (event: MouseEvent | TouchEvent): event is TouchEvent => {
-    return Boolean((event as TouchEvent).touches && (event as TouchEvent).touches.length);
-};
-export const isMouseEvent = (event: MouseEvent | TouchEvent): event is MouseEvent => {
-    return Boolean(
-        ((event as MouseEvent).clientX || (event as MouseEvent).clientX === 0) &&
-        ((event as MouseEvent).clientY || (event as MouseEvent).clientY === 0)
-    );
-};
